+++ backend/application/questions/dtos/SavedQuestions.ts
@@ -1,0 +1,6 @@
+import { Question } from '@/backend/domain/entities/Question';
+
+export interface SavedQuestionsResult {
+  questions: Question[];
+  reportId: number;
+}
