+++ backend/domain/dtos/QuestionsRequest.ts
@@ -1,0 +1,4 @@
+export interface QuestionsRequest {
+  buffer: Buffer;
+  fileName: string;
+}
