+++ backend/domain/dtos/QuestionsResponse.ts
@@ -1,0 +1,4 @@
+export interface QuestionsResponse {
+  order: number;
+  question: string;
+}
